pyproject.toml
[project]
name = "gh-issues-local"
version = "0.1.0"
description = "A local implementation of the GitHub Issues API built with FastAPI"
readme = "README.md"
authors = [
    { name = "DavidKoleczek", email = "45405824+DavidKoleczek@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.129.0",
    "pyyaml>=6.0",
    "storage-provider",
]

[project.scripts]
gh-issues-local = "gh_issues_local:main"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.15.0",
    "ty>=0.0.16",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = [
    "F",    # pyflakes
    "E",    # pycodestyle
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "RUF",  # ruff
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "ISC",  # flake8-implicit-str-concat
    "PIE",  # flake8-pie
    "PT",   # flake-pytest-style
    "PTH",  # flake8-use-pathlib
    "SIM",  # flake8-simplify
    "TID",  # flake8-tidy-imports
]
ignore = [
    "E501",   # line-too-long (formatter makes best-effort attempt)
    # The following rules conflict with the formatter, see: https://docs.astral.sh/ruff/formatter/#conflicting-lint-rules
    "W191",   # tab-indentation
    "E111",   # indentation-with-invalid-multiple
    "E114",   # indentation-with-invalid-multiple-comment
    "E117",   # over-indented
    "ISC002", # multi-line-implicit-string-concatenation
]

[tool.ruff.lint.isort]
force-sort-within-sections = true
split-on-trailing-comma = false

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.uv.sources]
storage-provider = { git = "https://github.com/DavidKoleczek/storage-provider.git" }
